+++ docs/NEWS.md
@@ -1,8 +1,13 @@
+# Changelog
+
 ## packer 0.0.7.9000
 
 - Ensure existing R files are not overwritten is already existing.
 - Added `make_library` to turn the project into a library 
 callable from outside the bundle.
+- Allow `@import` statements for CSS loader
+- Changed improved docs
+- Add `modules` and `import` arguments to CSS loader function.
 
 ## packer 0.0.6